+++ internal/agent/f5/services.go
@@ -75,8 +75,11 @@ func (a *Agent) getExtendedService(ctx context.Context, s *models.Service) (*as3
 		found := false
 		for _, ip := range service.IPAddresses {
 			var ipAddress net.IP
-			if ipAddress, _, err = net.ParseCIDR(ip.String()); err != nil {
-				return nil, fmt.Errorf("ParseCIDR: %w", err)
+			if ipAddress, _, err = net.ParseCIDR(string(ip)); err != nil {
+				ipAddress = net.ParseIP(string(ip))
+				if ipAddress == nil {
+					return nil, fmt.Errorf("ParseCIDR: invalid CIDR address: %s", ip)
+				}
 			}
 
 			if ipnet.Contains(ipAddress) {